Remove deprecated helper methods from SectorIndex
